+++ src/mito2/src/compaction/scheduler_test.rs
@@ -21,17 +21,24 @@ use common_datasource::compression::CompressionType;
 use common_meta::key::schema_name::SchemaNameValue;
 use common_time::{DatabaseTimeToLive, Timestamp};
 use store_api::storage::FileId;
-use tokio::sync::{Barrier, oneshot};
+use tokio::sync::{Barrier, mpsc, oneshot};
 
+use super::planning::CompactionPlanningResult;
+use super::state::{CompactingFiles, CompactionPhase};
 use super::*;
-use crate::compaction::CompactionOutput;
 use crate::compaction::memory_manager::{CompactionMemoryGuard, new_compaction_memory_manager};
+use crate::compaction::picker::PickerOutput;
 use crate::compaction::test_util::new_file_handle;
+use crate::compaction::{CompactionOutput, find_dynamic_options};
 use crate::error::InvalidSchedulerStateSnafu;
 use crate::manifest::manager::{RegionManifestManager, RegionManifestOptions};
+use crate::metrics::COMPACTION_MEMORY_REJECTED;
 use crate::region::ManifestContext;
+use crate::request::{BackgroundNotify, OutputTx, WorkerRequest};
+use crate::schedule::remote_job_scheduler::{RemoteJob, RemoteJobSchedulerRef};
 use crate::schedule::scheduler::{Job, Scheduler};
 use crate::sst::FormatType;
+use crate::sst::file::FileHandle;
 use crate::test_util::mock_schema_metadata_manager;
 use crate::test_util::scheduler_util::{SchedulerEnv, VecScheduler};
 use crate::test_util::version_util::{VersionControlBuilder, apply_edit};
